feat(core): main.jsからindex.htmlの初期化処理を分離
Issue #23 に基づき、アプリケーションのエントリーポイントである `main.js` をリファクタリング。

- `index.html` に固有の初期化ロジック（テーブル管理、日付ピッカー）を `indexPage.js` に分離。
- `main.js` は、ページのパスに応じて適切な初期化関数を呼び出す役割に徹するように修正。

これにより、`main.js` の責務が明確になり、コードの保守性と見通しが向上しました。
